add: lodash library
