+++ app/(site)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 
-import Footer from "@/_components/layout/Footer";
-import Header from "@/_components/layout/Header";
+import Footer from "@/components/layout/Footer";
+import Header from "@/components/layout/Header";
 
 import "@/styles/global.scss";
