+++ tsconfig.json
@@ -20,7 +20,6 @@
     ],
     "baseUrl": "./app",
     "paths": {
-      "@/*": ["*"],
       "@/components/*": ["_components/*"],
       "@/utils/*": ["_utils/*"],
       "@/styles/*": ["_styles/*"],
